Move additional data to types
