+++ src/components/Property/Types/OptionProperty.tsx
@@ -1,18 +1,7 @@
 import { Select } from '@mantine/core';
 
 import { useGetOptionPropertyValue, useGetPropertyDescription } from '@/api/hooks';
-import { PropertyProps } from '@/components/Property/types';
-
-// In OpenSpace the options are represented like so:
-// { 0: "Option 1"}, { 1: "Option 2"}
-// The key is a number but in a string format.
-interface Option {
-  [key: string]: string;
-}
-
-export interface AdditionalDataOptions {
-  Options: Option[];
-}
+import { AdditionalDataOptions, PropertyProps } from '@/components/Property/types';
 
 export function OptionProperty({ uri, readOnly }: PropertyProps) {
   const [value, setValue] = useGetOptionPropertyValue(uri);
