+++ src/components/Property/types.ts
@@ -12,3 +12,30 @@ export interface AdditionalDataVectorMatrix {
   MinimumValue: number[];
   SteppingValue: number[];
 }
+
+export interface AdditionalDataNumber {
+  Exponent: number;
+  MaximumValue: number;
+  MinimumValue: number;
+  SteppingValue: number;
+}
+
+export type ViewOptionsVector = {
+  Color?: boolean;
+  MinMaxRange?: boolean;
+};
+
+// In OpenSpace the options are represented like so:
+// { 0: "Option 1"}, { 1: "Option 2"}
+// The key is a number but in a string format.
+interface Option {
+  [key: string]: string;
+}
+
+export interface AdditionalDataOptions {
+  Options: Option[];
+}
+
+export interface AdditionalDataSelection {
+  Options: string[];
+}
